pyproject.toml
[tool.poetry]
name = "dlan-cast"
version = "0.1.0"
description = ""
authors = ["weihong.xu <xuweihong.cn@gmail.com>"]
packages = [{include = "dlan_cast"}]


[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.4.0"
upnpclient = "^1.0.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
